app/requirements.txt
asgiref==3.7.2
Django==4.2.2
sqlparse==0.4.4
tzdata==2023.3
django-environ==0.10.0
psycopg2-binary==2.9.6
djangorestframework==3.14.0
Markdown==3.4.3
django-filter==23.2
PyJWT==2.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
idna==3.4
oauthlib==3.2.2
pycparser==2.21
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
urllib3==2.0.3
annotated-types==0.5.0
pydantic==2.0.3
pydantic-core==2.3.0
typing-extensions==4.7.1
colorama==0.4.6
exceptiongroup==1.1.2
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytest-django==4.5.2
tomli==2.0.1
Faker==19.2.0
factory-boy==3.3.0
inflection==0.5.1
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
six==1.16.0